+++ NoiseHoiHoi/crates/noise-hoihoi-app/src/main.rs
@@ -1,6 +1,8 @@
 #![cfg_attr(windows, windows_subsystem = "windows")]
 
 #[cfg(windows)]
+mod signal_monitor;
+#[cfg(windows)]
 mod windows_app;
 
 #[cfg(windows)]
@@ -10,5 +12,5 @@ fn main() -> eframe::Result {
 
 #[cfg(not(windows))]
 fn main() {
-    eprintln!("NoiseHoiHoi v0.1 only supports Windows 11 x64");
+    eprintln!("NoiseHoiHoi currently supports Windows 11 x64");
 }